backend/requirements.txt
fastapi
langchain
ollama
pymupdf4llm
pydantic
python-dotenv
qdrant-client
uvicorn
